Action and subclasses converted to double

+++ src/include/constants.h
@@ -26,7 +26,7 @@ constexpr float fm2_mb = 0.1;
 /**
  * Numerical error tolerance
  */
-constexpr float really_small = 1.0e-6;
+constexpr double really_small = 1.0e-6;
 /**
  * \f$ 2\pi \f$
  *
@@ -47,25 +47,25 @@ constexpr float small_number = 1.0e-4;
  *
  * Note that this should be the same as in particles.txt.
  */
-constexpr float nucleon_mass = 0.938;
+constexpr double nucleon_mass = 0.938;
 /**
  * pion mass in GeV
  *
  * Note that this should be the same as in particles.txt.
  */
-constexpr float pion_mass = 0.138;
+constexpr double pion_mass = 0.138;
 /**
  * kaon mass in GeV
  *
  * Note that this should be the same as in particles.txt.
  */
-constexpr float kaon_mass = 0.494;
+constexpr double kaon_mass = 0.494;
 /**
  * Delta mass in GeV
  *
  * Note that this should be the same as in particles.txt.
  */
-constexpr float delta_mass = 1.232;
+constexpr double delta_mass = 1.232;
 
 /**
  * Fine-struture constant, approximately 1/137
